zeromq/jeromq wrapper bug fix

+++ messaging/jeromq_compat.py
@@ -2,11 +2,17 @@ import cPickle as pickle
 #import pickle
 from ZMQ_Again_Exception import ZMQ_Again
 
+USE_JNIUS = True
 
-from jnius import autoclass
-zmq = autoclass('org/zeromq/ZMQ')
-Context = autoclass('org.zeromq.ZContext')
-Subscriber = autoclass('com.ddw_gd.jeromqfixer.JeromqFixer')
+if USE_JNIUS:
+    from jnius import autoclass
+    zmq = autoclass('org/zeromq/ZMQ')
+    Context = autoclass('org.zeromq.ZContext')
+    Subscriber = autoclass('com.ddw_gd.jeromqfixer.JeromqFixer')
+else:
+    from org.zeromq import ZMQ as zmq
+    from org.zeromq import ZContext as Context
+    from com.ddw_gd.jeromqfixer import JeromqFixer as Subscriber
 
 
 class ZeroMQ:
@@ -18,6 +24,7 @@ class ZeroMQ:
     def __init__(self, address, mode):
         self.context = Context()
         self.socket = self.context.createSocket(mode)
+        self.closed = False
 
         if mode == self.REQ:
             self.socket.connect(address)
@@ -31,8 +38,10 @@ class ZeroMQ:
 
 
     def close(self):
-        self.socket.close()
-        self.context.destroy()
+        if not self.closed:
+            self.closed = True
+            self.socket.close()
+            self.context.destroy()
 
 
     def recv_pyobj(self, blocking=True):
